requirements.txt
pycryptodome
PyPDF2